Cleaning, seperated IOC into seperate folder
