+++ client/src/pages/EducatorDashboard/Individual Oppurtunity/EventSection.tsx
